+++ speechflow-cli/src/speechflow.ts
@@ -299,6 +299,7 @@ let debug = false
         "./speechflow-node-t2t-awstranslate.js",
         "./speechflow-node-t2t-deepl.js",
         "./speechflow-node-t2t-format.js",
+        "./speechflow-node-t2t-google.js",
         "./speechflow-node-t2t-ollama.js",
         "./speechflow-node-t2t-openai.js",
         "./speechflow-node-t2t-sentence.js",